add support for SCO fake sink suspend prevention
This commit brings the code from Canonical which make the sink,
depending on the output selected, prevent the suspension of the
Bluetooth headset. This is forward-ported from the old version using
the code from [1].

For this to work properly, it also requires a custom patch in the
Pulseaudio's bluez module.

[1]
  - https://github.com/saemy/pulseaudio/commit/4bb6142da20b7f38c2a343a4f007a3c0c2787b97
  - https://github.com/saemy/pulseaudio/commit/60d26f12dc67af7f634aa4e5f361e7562ec050b1